Consolidate into less packages
