+++ cmake/EkatCreateUnitTest.cmake
@@ -96,14 +96,11 @@ function(EkatCreateUnitTestExec exec_name exec_srcs)
   endforeach()
 
   # Link flags/libs
-  target_link_libraries(${target_name} PUBLIC TestUtils)
   if (NOT ecute_EXCLUDE_MAIN_CPP)
     target_link_libraries(${target_name} PUBLIC CatchMain)
   endif ()
   if (ecute_USER_DEFINED_TEST_SESSION)
     target_compile_options(${target_name} PRIVATE USER_DEFINED_TEST_SESSION)
-  else ()
-    target_link_libraries(${target_name} PUBLIC Session)
   endif ()
   if (ecute_LIBS)
     target_link_libraries(${target_name} PUBLIC "${ecute_LIBS}")
