+++ src/catch-main/ekat_catch_main.cpp
@@ -3,17 +3,17 @@
 #include "catch2/catch.hpp"
 
 #include "ekat_test_utils.hpp"
+#include "ekat_arch.hpp"
+#include "ekat_fpe.hpp"
 #include "ekat_assert.hpp"
 
-#ifdef EKAT_ENABLE_MPI_COMM
 #include "ekat_comm.hpp"
 #ifdef EKAT_ENABLE_MPI
 #include <mpi.h>
 #endif
-#endif
 
-#ifndef USER_DEFINED_TEST_SESSION
-#include "ekat_session.hpp"
+#if !defined(USER_DEFINED_TEST_SESSION) && defined(EKAT_HAS_KOKKOS)
+#include "ekat_kokkos_session.hpp"
 #endif
 
 // If USER_DEFINED_TEST_SESSION is defined, the user MUST provide implementations
@@ -70,31 +70,27 @@ int main (int argc, char **argv) {
   EKAT_REQUIRE_MSG (ret==0,"Error! Something went wrong while parsing command line.\n");
 
   bool am_i_root = true;
-#ifdef EKAT_ENABLE_MPI_COMM
   ekat::Comm comm;
 #ifdef EKAT_ENABLE_MPI
   comm.reset_mpi_comm(MPI_COMM_WORLD);
 #endif
   am_i_root = comm.am_i_root();
-#endif
 
-  // Initialize test session (initializes kokkos and print config settings).
-  // Ekat provides a default impl, but the user can choose
-  // to not use it, and provide one instead.
+  // Initialize the test session
+  // Ekat provides a default impl, but the user can choose to not use it, and provide one instead.
+  // The default impl initializes kokkos and print config settings.
   ekat_initialize_test_session(argc,argv,am_i_root);
 
-#ifdef EKAT_ENABLE_MPI_COMM
   comm.barrier();
   std::cout << "Starting catch session on rank " << comm.rank() << " out of " << comm.size() << "\n";
   comm.barrier();
-#endif
 
   // Run tests
   int num_failed = catch_session.run();
 
-  // Finalizes test session (finalizes kokkos).
-  // Ekat provides a defalt impl, but the user can choose
-  // to not use it, and provide one instead.
+  // Finalizes test session
+  // Ekat provides a defalt impl, but the user can choose to not use it, and provide one instead.
+  // The default impl finalizes kokkos
   ekat_finalize_test_session ();
 
 #ifdef EKAT_ENABLE_MPI
@@ -108,7 +104,9 @@ int main (int argc, char **argv) {
 
 #ifndef USER_DEFINED_TEST_SESSION
 void ekat_initialize_test_session (int argc, char** argv, const bool print_config) {
-  ekat::initialize_ekat_session (argc,argv,print_config);
+#ifdef EKAT_HAS_KOKKOS
+  ekat::initialize_kokkos_session (argc,argv,print_config);
+#endif
 
 #ifdef EKAT_ENABLE_FPE_DEFAULT_MASK
   int mask = FE_DIVBYZERO | FE_INVALID | FE_OVERFLOW;
@@ -116,13 +114,15 @@ void ekat_initialize_test_session (int argc, char** argv, const bool print_confi
 #endif
 
   if (print_config) {
-    printf(" active avx set: %s\n", active_avx_string().c_str());
-    printf(" compiler id: %s\n", compiler_id_string().c_str());
-    printf(" current FPE: mask=%d, %s\n", get_enabled_fpes(), fpe_config_string().c_str());
+    printf(" active avx set: %s\n", ekat::active_avx_string().c_str());
+    printf(" compiler id: %s\n", ekat::compiler_id_string().c_str());
+    printf(" current FPE: mask=%d, %s\n", ekat::get_enabled_fpes(), ekat::fpe_config_string().c_str());
   }
 }
 
 void ekat_finalize_test_session () {
-  ekat::finalize_ekat_session ();
-}
+#ifdef EKAT_HAS_KOKKOS
+  ekat::finalize_kokkos_session ();
 #endif
+}
+#endif // !defined(USER_DEFINED_TEST_SESSION)
