+++ src/core/ekat_rational_constant.hpp
