+++ src/core/ekat_string_utils.cpp
