+++ tests/CMakeLists.txt
@@ -82,9 +82,6 @@ endif()
 # Core is ALWAYS enabled
 add_subdirectory(core)
 
-# And test-utils IS required for testing
-add_subdirectory(test-utils)
-
 # Process sub-pacakges as requested
 if (EKAT_ENABLE_ALGORITHM)
   add_subdirectory(algorithm)
@@ -92,21 +89,9 @@ endif()
 if (EKAT_ENABLE_KOKKOS)
   add_subdirectory(kokkos)
 endif()
-if (EKAT_ENABLE_MPI_COMM)
-  add_subdirectory(mpi)
-endif()
 if (EKAT_ENABLE_PACK)
   add_subdirectory(pack)
 endif()
 if (EKAT_ENABLE_YAML_PARSER)
-  add_subdirectory(yaml)
-endif()
-if (EKAT_ENABLE_TESTING)
-  add_subdirectory(testing)
-endif()
-if (EKAT_ENABLE_UNITS)
-  add_subdirectory(units)
-endif()
-if (EKAT_ENABLE_UTILS)
-  add_subdirectory(utils)
+  add_subdirectory(parser)
 endif()
