+++ tests/algorithm/tridiag_tests.cpp
@@ -3,11 +3,11 @@
 
 #include "ekat_test_config.h"
 #include "tridiag_tests.hpp"
-#include "ekat_session.hpp"
+#include "ekat_kokkos_session.hpp"
 
 int main (int argc, char **argv) {
   int num_failed = 0;
-  ekat::initialize_ekat_session(argc, argv); {
+  ekat::initialize_kokkos_session(argc, argv); {
     if (argc > 1) {
       // Performance test.
       ekat::test::perf::Input in;
@@ -20,6 +20,6 @@ int main (int argc, char **argv) {
       // Correctness tests.
       num_failed = Catch::Session().run(argc, argv);
     }
-  } ekat::finalize_ekat_session();
+  } ekat::finalize_kokkos_session();
   return num_failed != 0 ? 1 : 0;
 }
