+++ tests/core/catch_main_tests.cpp
