+++ tests/core/factory.cpp
