+++ tests/core/std_meta.cpp
