+++ tests/core/string_utils.cpp
