--- tests/units/CMakeLists.txt
@@ -1,4 +1,0 @@
-include(EkatCreateUnitTest)
-
-# Test units framework
-EkatCreateUnitTest(units "units.cpp" LIBS Units)
